requirements.txt
torch
transformers
tensorboard
opencv-python
scikit-learn
pandas
numpy
matplotlib
seaborn
Pillow
streamlit
pyyaml
